python_sql/requirements.txt
astroid==2.15.5
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
dill==0.3.6
executing==1.2.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mysql-connector-python==8.0.33
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pylint==2.17.4
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
wrapt==1.15.0
